Fix: Adjust Spots migration to use floating point and defined decimal integers

+++ apps/backend/db/migrations/20240812031753-create-spot.js
@@ -37,11 +37,11 @@ module.exports = {
 					allowNull: false,
 				},
 				lat: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				lng: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				name: {
@@ -53,7 +53,7 @@ module.exports = {
 					allowNull: false,
 				},
 				price: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.DECIMAL(10, 2),
 					allowNull: false,
 				},
 				createdAt: {
